Extracted EditableDesc and refactored it's usage.
