+++ domains/eventEditor/src/ui/datetimes/datesList/editable/index.ts
@@ -1,3 +1,1 @@
-export { default as EditableDesc } from './EditableDesc';
-
 export { default as EditableName } from './EditableName';
