+++ domains/eventEditor/src/ui/shared/editable/index.ts
@@ -1,0 +1,1 @@
+export { default as EditableDesc } from './EditableDesc';
