+++ domains/eventEditor/src/ui/tickets/ticketsList/editable/index.ts
@@ -1,5 +1,3 @@
-export { default as EditableDesc } from './EditableDesc';
-
 export { default as EditableName } from './EditableName';
 
 export { default as EditablePrice } from './EditablePrice';
