+++ packages/components/src/TabbableText/types.ts
@@ -1,0 +1,8 @@
+export interface TabbableTextProps {
+	className?: string;
+	icon?: React.ReactNode;
+	onClick: VoidFunction;
+	richTextContent?: boolean;
+	text?: string | JSX.Element;
+	tooltip?: string;
+}
