+++ packages/components/src/index.ts
@@ -41,6 +41,7 @@ export * from './SearchInput';
 export * from './SelectInput';
 export * from './Stepper';
 export * from './SwitchInput';
+export * from './TabbableText';
 export * from './TimezoneTimeInfo';
 export * from './withLabel';
 export * from './withTooltip';
